UI: Rename Distort socket in the Lens Distortion compositing node
The "Distort" socket in the Lens Distortion node should be a noun
instead of a verb, same as "Dispersion" just below. The name was
introduced at the same time as the node itself, in 2a2453d3e2.

Versioning was added to update existing nodes. The Python API
is affected though, if the socket was looked up with the old name.

Pull Request: https://projects.blender.org/blender/blender/pulls/108234